(major) remove default pandoc templating
The default pandoc template seemed to be messing with the way code
blocks are indented (why is this so?), so it had to be removed. For now,
code blocks are indented just as they are in fenced code blocks in the
markdown source.

TODO: test why the default template was screwing with the code blocks.
If safe, remove the templates folder.

- make top-bar links more friendly to keyboard-only navigation.
  Essentially swapped "Up one Page" and "Back to ToC".
- added basic stuff in ML notes.