(major) standardize build process with Makefile
Markdown -> HTML conversion is now done via a Makefile, which ensures
that no unnecessary conversions are performed (i.e. with unchanged
files).
Additionally, to aid organization, all source markdown files live in
src/ and all generated HTML lives in docs/ (this folder is so named because
of GitHub Pages).

TODO: check if Jekyll can be configured to look into a different root
path for the page.