(major) choose template from tajmone/pandoc-goodies
The default pandoc template for some reason screwes up the
indentation in code blocks (see 93da9de). Testing reveals that this
probably has to do with interference between default css styles (part of
default.html) and the minimal.css used.

What's weirder is that this happens only when the option
--template=default.html is used and NOT while using --standalone (even
though they're supposed to be the same?).

Hence, to save time, move to the HTML5 Github-ish pandoc template from
tajmone/pandoc-goodies, which seems to work just fine with the css
styling removed.